+++ src/layouts/AdminLayout.tsx
@@ -1,7 +1,9 @@
+import { useState } from 'react'
 import { Outlet } from 'react-router-dom'
 import { Calendar, Users, Settings } from 'lucide-react'
-import { Sidebar } from './Sidebar'
-import { PATHS } from '../../../routes/paths'
+import { Sidebar } from '../shared/components/layout/Sidebar'
+import { LogoutModal } from '../features/auth/components/LogoutModal'
+import { PATHS } from '../routes/paths'
 
 const adminNavItems = [
   { to: PATHS.ADMIN_SHIFTS, label: 'シフト管理', icon: Calendar },
@@ -10,12 +12,15 @@ const adminNavItems = [
 ]
 
 export function AdminLayout() {
+  const [logoutOpen, setLogoutOpen] = useState(false)
+
   return (
     <div className="flex h-screen overflow-hidden bg-gray-50">
-      <Sidebar items={adminNavItems} />
+      <Sidebar items={adminNavItems} onLogoutClick={() => setLogoutOpen(true)} />
       <main className="flex-1 overflow-auto">
         <Outlet />
       </main>
+      <LogoutModal open={logoutOpen} onClose={() => setLogoutOpen(false)} />
     </div>
   )
 }
