+++ src/layouts/AuthLayout.tsx
