+++ src/routes/config.tsx
@@ -8,8 +8,8 @@ import { LoginPage } from '../pages/LoginPage'
 import { NotFoundPage } from '../pages/NotFoundPage'
 import { ResetPasswordPage } from '../pages/ResetPasswordPage'
 import { SettingsPage } from '../pages/SettingsPage'
-import { AdminLayout } from '../shared/components/layout/AdminLayout'
-import { AuthLayout } from '../shared/components/layout/AuthLayout'
+import { AdminLayout } from '../layouts/AdminLayout'
+import { AuthLayout } from '../layouts/AuthLayout'
 import { RouteGuard } from './RouteGuard'
 import { PATHS } from './paths'
